Memperbaiki name case pada file dan React Component
